Finalizado o ajuste, do cadastramento do produto e o retorno do mesmo.
Iniciado a implementação do Pessoa, sendo fisica e jurica.

+++ src/main/java/onecenter/com/br/ecommerce/controller/produtos/ProdutosController.java
@@ -22,7 +22,7 @@ public class ProdutosController {
     @PostMapping(value = "/criar", consumes = {"multipart/form-data"})
     public ResponseEntity<ProdutosResponse> criarProduto(
             @RequestParam("nome") String nome,
-            @RequestParam("preco") Float preco,
+            @RequestParam("preco") Double preco,
             @RequestParam("id_categoria") Integer idCategoria,
             @RequestParam("produto_imagem") MultipartFile  imagem) {
