+++ src/main/java/onecenter/com/br/ecommerce/dto/produtos/request/ProdutoRequest.java
@@ -14,7 +14,7 @@ public class ProdutoRequest {
 
     private Integer id_produto;
     private String nome;
-    private Float preco;
+    private Double preco;
     private MultipartFile produto_imagem;
     private Integer id_categoria;
 }
